[Lexer] add comment module and test
